Refactor: Moved connection strings to AWS Secrets instead of pulling individual parameters from there' '
